:bug: fix: viewer model load
